% Generated by roxygen2 (4.0.1): do not edit by hand
\name{praat}
\alias{praat}
\title{Execute a Praat command}
\usage{
praat(command, arguments = list(), input = NULL, output = NULL,
  overwrite = NULL, filetype = "text", simplify = FALSE)
}
\arguments{
\item{command}{A character string indicating what command should be executed.
This will be checked against the 'SupportedCommands' database, and if it's
not found there, an error message will be issued.}

\item{arguments}{A list() of arguments. This data structure is used because
it is possible to mix of various classes, e.g. list("character", 12.34, 5,
TRUE).}

\item{input}{The (ideally full) path to the input file.}

\item{output}{The (ideally full) path to the output file.}

\item{overwrite}{Only applicable if the user chooses a 'Create' or 'Modify'
command. (If they choose a 'Query' command, this will be ignored.) This
indicates whether the command should be executed if the file specified in
'output' already exists (thereby overwriting it). By defaulting to FALSE,
the user is protected from accidentally erasing data.}

\item{filetype}{Determines file format for the output text file. Only
applicable if the user chooses a 'Create' or 'Modify' command. Three
choices: "text" = Save as text file..., "short" = Save as short text
file..., and "binary" = Save as binary file...}

\item{simplify}{Only applicable if the user chooses a 'Query' command.
Determines whether the queried information will be coerced to numeric (i.e.
just the core data) rather than remain as a text string (potentially along
with various kinds of embellishing explanatory information). Either way,
the result will (initially) be brought into R as a character string. The
user specifies this as TRUE or FALSE. This later gets translated into 1 or
0 when passed to Praat.}
}
\value{
(place-holder text)
}
\description{
Execute a Praat command
}

